+++ examples/example2.py
@@ -3,25 +3,14 @@ import os
 #os.getcwd() should change to pyautocv if not
 os.chdir("pyautocv")
 from pyautocv.segmentation import *
-to_smooth = EdgeDetection("images/people","sobel_vertical")
-show_images(to_smooth.gray_images(), to_smooth.smooth())
-# detect edges
-edge_detection = EdgeDetection("images","sobel_vertical")
-show_images(edge_detection.read_images(), edge_detection.detect_edges(operator="laplace",mask="gaussian",sigma=3.5))
-
-# Thresholding
-to_threshold = Threshold("images/biology",threshold_method="binary")
+edge_detection = Segmentation("images/cats")
+show_images(edge_detection.read_images(), edge_detection.detect_edges(operator="laplace", mask="gaussian", sigma=0))
+# Threshold
+to_threshold = Segmentation("images/biology")
 show_images(to_threshold.read_images(),to_threshold.threshold_images())
-# cats
-to_threshold = Threshold("images/cats",threshold_method="binary")
+# threshold otsu
+to_threshold = Segmentation("images/biology")
 show_images(to_threshold.read_images(),to_threshold.threshold_images(threshold_method="otsu"))
-# potholes
-to_threshold = Threshold("images/potholes",threshold_method="otsu")
-show_images(to_threshold.read_images(),to_threshold.threshold_images())
-# houses
-to_threshold = Threshold("images/houses",threshold_method="binary_inverse")
-show_images(to_threshold.read_images(),to_threshold.threshold_images())
-
-# people
-to_threshold = Threshold("images/people",threshold_method="binary_inverse")
-show_images(to_threshold.read_images(),to_threshold.threshold_images())
+# cats
+to_threshold = Segmentation("images/cats")
+show_images(to_threshold.read_images(),to_threshold.threshold_images(threshold_method="binary_inverse"))