+++ pyautocv/segmentation.py
@@ -3,10 +3,21 @@ import cv2
 import matplotlib.pyplot as plt
 from scipy import ndimage
 from skimage.io import ImageCollection
+from skimage import filters
 from os import pathsep
 import numpy as np
 from itertools import chain
 
+
+
+def gray_images(images):
+    """
+
+    :return: Returns grayed images. Currently supports only blue-green-red to gray conversion
+
+    """
+
+    return list(map(lambda x: cv2.cvtColor(x, cv2.COLOR_BGR2GRAY), images))
 
 # init class segmentation
 class Segmentation(object):
@@ -18,6 +29,8 @@ class Segmentation(object):
 
         """
         self.directory = directory
+        if self.directory is None:
+            self.directory = "."
 
     def read_images(self):
         """
@@ -32,15 +45,8 @@ class Segmentation(object):
 
         return list(images_list)
 
-    def gray_images(self):
-        """
 
-        :return: Returns grayed images. Currently supports only blue-green-red to gray conversion
-
-        """
-        return list(map(lambda x: cv2.cvtColor(x, cv2.COLOR_BGR2GRAY), self.read_images()))
-
-    def smooth(self, mask="box", sigma=3, kernel_shape=(3, 3)):
+    def smooth(self, mask="box",kernel_shape=(5, 5), **kwargs):
         """
 
         :param sigma: Sigma to use for a gaussian filter
@@ -50,22 +56,16 @@ class Segmentation(object):
 
         """
         image_list = self.read_images()
-        mask_list = {'gaussian': lambda x: ndimage.gaussian_filter(x, sigma=sigma),
+        mask_list = {'gaussian': lambda x: ndimage.gaussian_filter(x, **kwargs),
                      'box': lambda x: cv2.blur(x, ksize=kernel_shape),
-                     'median': lambda x: cv2.medianBlur(x, ksize=kernel_shape)}
+                     'median': lambda x: cv2.medianBlur(x, ksize=kernel_shape[0])}
         # alias box with mean, would be great to have a .alias method
-        mask_list.update({"mean": mask_list['box']})
+        mask_list.update({"mean": mask})
         # convolve images with low pass filter
+        print("Smoothing with {}".format(mask))
         low_pass_filtered = list(map(mask_list[mask], image_list))
 
         return low_pass_filtered
-
-
-class Threshold(Segmentation):
-    def __init__(self, directory, threshold_method):
-        super().__init__(directory)
-        self.directory = directory
-        self.threshold_method = threshold_method
 
     def threshold_images(self, threshold_method="binary", use_threshold=127, use_max=255):
         """
@@ -79,7 +79,6 @@ class Threshold(Segmentation):
         """
         # use cv2's threshold instead since this is more mature, don't reinvent
         # There must be a different way than a list-map approach
-        self.threshold_method = threshold_method
         threshold_methods = {'binary': lambda x: cv2.threshold(x, use_threshold, use_max, cv2.THRESH_BINARY),
                              'binary_inverse': lambda x: cv2.threshold(x, use_threshold, use_max,
                                                                        cv2.THRESH_BINARY_INV),
@@ -87,7 +86,7 @@ class Threshold(Segmentation):
                                                                        cv2.THRESH_TOZERO),
                              'otsu': lambda x: cv2.threshold(x, use_threshold, use_max,
                                                              cv2.THRESH_BINARY + cv2.THRESH_OTSU)}
-        image_list = self.gray_images()
+        image_list = gray_images(self.read_images())
 
         thresholded_images = list(map(threshold_methods[threshold_method], image_list))
         # drop ret val
@@ -96,21 +95,10 @@ class Threshold(Segmentation):
         return thresholded_images
 
 
-class EdgeDetection(Segmentation):
 
-    def __init__(self, directory, operator):
-        super().__init__(directory)
-        self.directory = directory
-        self.operator = operator
 
-    # need
-    def available_operators(self):
-        available_operators = ["sobel_horizontal", "sobel_vertical", "prewitt_horizontal", "prewitt_vertical",
-                               "laplace", "roberts_vertical", "roberts_horizontal", "scharr_vertical",
-                               "scharr_horizontal"]
-        return available_operators
 
-    def detect_edges(self, operator="sobel_vertical", kernel_size=3, **kwargs):
+    def detect_edges(self, operator="sobel_vertical", kernel_size=3,optional_mask=None,**kwargs):
         """
 
         :param kernel_size: int size to use for edge detection kernels
@@ -124,29 +112,28 @@ class EdgeDetection(Segmentation):
         self.operator = operator
         self.kernel_size = kernel_size
 
-        if operator not in self.available_operators():
-            raise ValueError("operator should be one of {}".format(self.available_operators()))
+        available_operators = ["sobel_horizontal", "sobel_vertical", "prewitt_horizontal", "prewitt_vertical",
+                               "laplace", "roberts_cross_neg", "roberts_horizontal", "scharr_vertical",
+                               "scharr_horizontal","canny", "roberts"]
+
+        if operator not in available_operators:
+            raise KeyError("operator should be one of {}".format(available_operators))
 
         kernels = {'sobel_horizontal': lambda x: cv2.Sobel(x, cv2.CV_64F, 1, 0, ksize=self.kernel_size),
                    'sobel_vertical': lambda x: cv2.Sobel(x, cv2.CV_64F, 0, 1, ksize=self.kernel_size),
-                   'roberts_horizontal': lambda x: ndimage.convolve(x, np.array([[1, 0], [0, -1]]), mode="reflect"),
-                   'roberts_vertical': lambda x: ndimage.convolve(x, np.array([[0, 1], [-1, 0]]), mode="reflect"),
+                   'roberts': lambda x: filters.roberts(x, optional_mask),
+                   'roberts_cross_neg': lambda x: lambda x: ndimage.convolve(x, np.array([[0, -1], [1, 0]])),
+                   'roberts_cross_pos': lambda x: filters.roberts_pos_diag(x, optional_mask),
                    'laplace': lambda x: cv2.Sobel(x, cv2.CV_64F, 1, 0, ksize=self.kernel_size),
-                   'prewitt_horizontal': lambda x: ndimage.convolve(x, np.array([[1, 0, -1], [1, 0, -1], [1, 0, -1]]),
-                                                                    mode="reflect"),
-                   'prewitt_vertical': lambda x: ndimage.convolve(x, np.array([[1, 1, 1], [0, 0, 0],
-                                                                               [-1, -1, -1]]),
-                                                                  mode="refelct"),
-                   'scharr_horizontal': lambda x: ndimage.convolve(x,
-                                                                   np.array([[3, 0, -3], [10, 0, -10], [3, 0, -3]]),
-                                                                   mode="reflect"),
-                   'scharr_vertical': lambda x: ndimage.convolve(x,
-                                                                 np.array([[3, 10, 3], [0, 0, 0], [-3, -10, -3]]),
-                                                                 mode="reflect")}
+                   'prewitt_horizontal': lambda x: filters.prewitt_h(x, optional_mask),
+                   'prewitt_vertical': lambda x: filters.prewitt_v(x, optional_mask),
+                   'scharr_horizontal': lambda x: filters.scharr_h(x, optional_mask),
+                   'scharr_vertical': lambda x: filters.scharr_v(x, optional_mask)}
 
         print("Using {}".format(self.operator))
-
-        final_images = list(map(kernels[self.operator], self.smooth(**kwargs)))
+        # denoise and gray
+        grayed_denoised = gray_images(self.smooth(**kwargs))
+        final_images = list(map(kernels[self.operator], grayed_denoised))
 
         return final_images
 
@@ -161,6 +148,7 @@ def show_images(original_images=None, processed_images=None):
         raise ValueError("Both original and processed image lists are required.")
 
     image_list = list(chain(original_images, processed_images))
+
 
     if len(image_list) % 2 == 0:
         ncols = len(image_list) / 2