+++ docs/source/changelog.md
@@ -1,5 +1,16 @@
-# Changes to pyautocv
+**pyautocv 0.2.1**
 
+**Major Changes**
+
+* Classes `Threshold` and `EdgeDetection` were dropped. Support is now through `Segmentation`
+
+* `gray_images` is now a static method. 
+
+**Additions**
+
+* Added support for more detection methods eg Roberts Cross
+
+----
 **pyautocv 0.2.0**
 
 **Major Additions**